+++ ionoscloud/test_constants.go
@@ -1808,6 +1808,7 @@ resource ` + constant.LoggingPipelineResource + ` ` + constant.LoggingPipelineTe
 const clusterIdAttribute = "cluster_id"
 const clusterInstancesAttribute = "instances"
 const clusterCoresAttribute = "cores"
+const clusterLocationAttribute = "location" // nolint:unused
 const clusterRamAttribute = "ram"
 const clusterStorageSizeAttribute = "storage_size"
 const clusterConnectionsAttribute = "connections"